- Added 'Role' field to user entity, created migration; - Added 'Roles' enum, 'Role' decorator and 'Roles Guard'; - Created 'CurrentUserId' decorator; - fixed paths in imports; - added logic which will create user with role 'Parent' during registration; - created method 'checkTaskOwnership' to check if task refers to current user;

+++ src/tasks/tasks.controller.ts
@@ -1,4 +1,4 @@
-import { BadRequestException, Body, Controller, Delete, Get, HttpCode, HttpStatus, NotFoundException, Param, Patch, Post, Query } from '@nestjs/common';
+import { BadRequestException, Body, Controller, Delete, ForbiddenException, Get, HttpCode, HttpStatus, NotFoundException, Param, Patch, Post, Query} from '@nestjs/common';
 import { TasksService } from './tasks.service';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { FindOneParams } from './find-one.params';
@@ -7,8 +7,9 @@ import { WrongTaskStatusException } from './exceptions/wrong-task-status.excepti
 import { Task } from './task.entity';
 import { CreateTaskLabelDto } from './dto/create-task-label.dto';
 import { FindTaskParams } from './find-task.params';
-import { PaginationParams } from 'src/common/pagination.params';
-import { PaginationResponse } from 'src/common/pagination.response';
+import { PaginationParams } from './../common/pagination.params';
+import { PaginationResponse } from './../common/pagination.response';
+import { CurrentUserId } from './../users/decorators/current-user-id.decorator';
 
 @Controller('tasks')
 export class TasksController {
@@ -18,10 +19,11 @@ export class TasksController {
     @Get()
     public async findAll(
         @Query() filters: FindTaskParams,
-        @Query() pagination: PaginationParams
+        @Query() pagination: PaginationParams,
+        @CurrentUserId() userId: string,
     ): Promise<PaginationResponse<Task>>
     {
-        const [items, total] = await this.tasksService.findAll(filters, pagination);
+        const [items, total] = await this.tasksService.findAll(filters, pagination,userId);
 
         return {
             data: items,
@@ -34,23 +36,34 @@ export class TasksController {
     };
 
     @Get('/:id')
-    public async findOne(@Param() params: FindOneParams): Promise<Task>
+    public async findOne(
+        @Param() params: FindOneParams,
+        @CurrentUserId() userId: string,
+    ): Promise<Task>
     {
-        return await this.findOneOrFail(params.id);
+        const task = await this.findOneOrFail(params.id);
+        this.checkTaskOwnership(task, userId);
+        return task;
     };
 
     @Post()
-    public async create(@Body() createTaskDto: CreateTaskDto): Promise<Task>
+    public async create(
+        @Body() createTaskDto: CreateTaskDto,
+        @CurrentUserId() userId: string,
+    ): Promise<Task>
     {
-        return await this.tasksService.createTask(createTaskDto);
+        return await this.tasksService.createTask({...createTaskDto, userId: userId});
     };
 
     @Patch('/:id')
     public async updateTask(
         @Param() params: FindOneParams,
-        @Body() updateTaskDto: UpdateTaskDto): Promise<Task>
+        @Body() updateTaskDto: UpdateTaskDto,
+        @CurrentUserId() userId: string,
+    ): Promise<Task>
     {
         const task = await this.findOneOrFail(params.id);
+        this.checkTaskOwnership(task, userId);
 
         try
         {
@@ -67,19 +80,25 @@ export class TasksController {
 
     @Delete('/:id')
     @HttpCode(HttpStatus.NO_CONTENT)
-    public async delete(@Param() params: FindOneParams): Promise<void>
+    public async delete(
+        @Param() params: FindOneParams,
+        @CurrentUserId() userId: string,
+    ): Promise<void>
     {
         const task = await this.findOneOrFail(params.id);
+        this.checkTaskOwnership(task, userId);
         await this.tasksService.deleteTask(task);
     };
 
     @Post(':id/labels')
     async addLabels(
         @Param() { id }: FindOneParams,
-        @Body() createTaskLabelDto: CreateTaskLabelDto[]): Promise<Task> 
+        @Body() createTaskLabelDto: CreateTaskLabelDto[],
+        @CurrentUserId() userId: string,
+    ): Promise<Task> 
     {
         const task = await this.findOneOrFail(id);
-
+        this.checkTaskOwnership(task, userId);
         return await this.tasksService.addLabels(task, createTaskLabelDto);
     };
 
@@ -87,10 +106,12 @@ export class TasksController {
     @HttpCode(HttpStatus.NO_CONTENT)
     async removeLabels(
         @Param() { id }: FindOneParams,
-        @Body() labelsNames: string[]): Promise<void>
+        @Body() labelsNames: string[],
+        @CurrentUserId() userId: string,
+    ): Promise<void>
     {
         const task = await this.findOneOrFail(id);
-
+        this.checkTaskOwnership(task, userId);
         await this.tasksService.removeLabels(task, labelsNames);
     };
 
@@ -105,5 +126,13 @@ export class TasksController {
         }
 
         return task;
+    };
+
+    private checkTaskOwnership(task: Task, userId: string): void
+    {
+        if (task.userId !== userId)
+        {
+        throw new ForbiddenException('You can only access your tasks');
+        }
     };
 }
